+++ src/Integration.php
@@ -296,7 +296,13 @@ class Integration extends AbstractGatewayIntegration {
 		$attempt = (int) $payment->get_meta( 'mollie_create_payment_attempt' );
 
 		if ( $attempt > 4 ) {
-			throw new \Exception( \sprintf( 'Could not create Mollie payment for %s after %s attempts.', $payment_id, $attempt ) );
+			throw new \Exception(
+				\sprintf(
+					'Could not create Mollie payment for %s after %s attempts.',
+					\esc_html( (string) $payment_id ),
+					\esc_html( (string) $attempt )
+				)
+			);
 		}
 
 		// Start payment.
@@ -442,7 +448,7 @@ class Integration extends AbstractGatewayIntegration {
 
 	/**
 	 * Payment fulfilled.
-	 * 
+	 *
 	 * @param Payment $payment Payment.
 	 * @return void
 	 */