+++ src/RefundTransformer.php
@@ -42,8 +42,8 @@ class RefundTransformer {
 	/**
 	 * Update Pronamic refund from Mollie refund.
 	 *
-	 * @param MollieRefund    $mollie_refund    Mollie refund.
-	 * @param PronamicPayment $pronamic_payment Pronamic payment.
+	 * @param MollieRefund   $mollie_refund   Mollie refund.
+	 * @param PronamicRefund $pronamic_refund Pronamic refund.
 	 * @return PronamicRefund
 	 */
 	public function update_mollie_to_pronamic( MollieRefund $mollie_refund, PronamicRefund $pronamic_refund ): PronamicRefund {
@@ -81,7 +81,7 @@ class RefundTransformer {
 
 		if ( null !== $mollie_refund->lines ) {
 			foreach ( $mollie_refund->lines as $mollie_line ) {
-				$id = $mollie_line->get_id();
+				$id = (string) $mollie_line->get_id();
 
 				if ( \array_key_exists( $id, $map_refund_lines ) ) {
 					$pronamic_refund_line = $map_refund_lines[ $id ];