+++ views/page-customers.php
@@ -8,6 +8,10 @@
  * @package   Pronamic\WordPress\Pay\Gateways\Mollie
  */
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
 ?>
 <div class="wrap">
 	<h1><?php echo \esc_html( \get_admin_page_title() ); ?></h1>