+++ CHANGELOG.md
@@ -7,6 +7,107 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [4.8.1] - 2023-10-16
+
+### Commits
+
+- Allow bypassing `first` sequence type with empty string (pronamic/wp-pronamic-pay-woocommerce#58). ([2a08130](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/2a08130afd975f1396d6ac213b0a7cb8e18496e6))
+
+Full set of changes: [`4.8.0...4.8.1`][4.8.1]
+
+[4.8.1]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.8.0...v4.8.1
+
+## [4.8.0] - 2023-10-13
+
+### Commits
+
+- Merge pull request #39 from pronamic/wp-mollie-4-dynamic-properties ([db6e935](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/db6e935910d7077c158f5817ced143c506716cc5))
+- Removed failure reason comment. ([83f2747](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/83f2747bbfbdb49b4d7d4917ddabeffe22a4201f))
+- Updated for payment details through `ObjectAccess` (pronamic/wp-mollie#4). ([450aca6](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/450aca61ca4a3e5efaffb195c249b3554a45c4f1))
+- Fixed the "The method parameter $args is never used" warnings. ([918d67d](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/918d67d6cac1fd7e72fd122e2378718ec4d477ff))
+- Fixed the "The method parameter $args is never used" warnings. ([1e3b3c7](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/1e3b3c73eabdc50aff2d0d6c7a8d5c79008f9fc6))
+- Cast vars to strings to make PHPStan happy. ([9fdee84](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/9fdee84e7f3cd209d273ed0afff1e05ab92e0178))
+- Compare meta value and Mollie ID as binary strings to avoid collate issues. ([fd373d1](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/fd373d18ab2ee08567b75042d2c72da4121d0ed6))
+- Require PHP >=8.0, updated pronamic/wp-coding-standards. ([d0daf64](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/d0daf64769c1abc2dce62f3c381df52190a612b4))
+
+### Composer
+
+- Changed `php` from `>=7.4` to `>=8.0`.
+- Changed `pronamic/wp-mollie` from `^1.2` to `v1.4.0`.
+	Release notes: https://github.com/pronamic/wp-mollie/releases/tag/v1.4.0
+- Changed `woocommerce/action-scheduler` from `^3.4` to `3.6.4`.
+	Release notes: https://github.com/woocommerce/action-scheduler/releases/tag/3.6.4
+- Changed `wp-pay/core` from `^4.9` to `v4.13.0`.
+	Release notes: https://github.com/pronamic/wp-pay-core/releases/tag/v4.13.0
+
+Full set of changes: [`4.7.11...4.8.0`][4.8.0]
+
+[4.8.0]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.7.11...v4.8.0
+
+## [4.7.11] - 2023-09-11
+
+### Commits
+
+- Make consumer name and IBAN fields required (https://github.com/pronamic/wp-pronamic-pay/issues/361). ([5f6d13c](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/5f6d13c32745fba007fd1765ec964f8014c1b474))
+- Added `wp-slug` to Composer config. ([a9f0539](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/a9f05394943165dde9496e7f371394ddbed45886))
+
+Full set of changes: [`4.7.10...4.7.11`][4.7.11]
+
+[4.7.11]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.7.10...v4.7.11
+
+## [4.7.10] - 2023-08-30
+
+### Fixed
+
+- Fixed setting Billie payment method status.
+
+### Commits
+
+- All payment methods are inactive by default. ([d15c1cc](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/d15c1cc82b5cf977c7a409d7cb161981aad03d8e))
+
+Full set of changes: [`4.7.9...4.7.10`][4.7.10]
+
+[4.7.10]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.7.9...v4.7.10
+
+## [4.7.9] - 2023-08-23
+
+### Commits
+
+- Fixed WPCS 3 issues. ([12b62fb](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/12b62fb3f733eb4a57c3a05f903f7273691f6fb7))
+- Dont allow direct file access. ([48c927b](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/48c927b97f30e9ff8b87d3b346589ac849001aa4))
+- Removed collate clause. ([bccdb0e](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/bccdb0eb6f5bc9ece8bd1146068758cb62a02270))
+
+Full set of changes: [`4.7.8...4.7.9`][4.7.9]
+
+[4.7.9]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.7.8...v4.7.9
+
+## [4.7.8] - 2023-07-12
+
+### Commits
+
+- Mark payment methods recurring support. ([a0c1246](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/a0c1246a55f5393c68c6d3a5e0f182a2d70eac84))
+- Added link to payment mandate ID. ([9d37959](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/9d37959c0c0baa92682136a0de9eb97be8b554a1))
+- Change incomplete `<dl>` elements to `<p>`. ([7f18457](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/7f18457ea5cbca1d3ef224dd0c0ac123392d9380))
+- Added support for Billie. ([59ea0a8](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/59ea0a8911cbe5521c4102343c53e58963ced6ab))
+- Added mandate page and link. ([277adc0](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/277adc0193c8faa6f56564d95716915b4fada334))
+
+Full set of changes: [`4.7.7...4.7.8`][4.7.8]
+
+[4.7.8]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.7.7...v4.7.8
+
+## [4.7.7] - 2023-06-01
+
+### Commits
+
+- Switch from `pronamic/wp-deployer` to `pronamic/pronamic-cli`. ([1659c0a](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/1659c0ac9a0713fef7de66eeca72ff9d9a2afde8))
+- Prevent error when trying to retrieve iDEAL issuers when SEPA Direct Debit and iDEAL are both inactive. ([1f3e4a0](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/1f3e4a0ea58c60d0ae268cb70774861962acfb16))
+- Fixed method documentation. ([949cf76](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/949cf76e561f176e5264f94f16f620ee89df03a7))
+- Updated .gitattributes ([45c4a5c](https://github.com/pronamic/wp-pronamic-pay-mollie/commit/45c4a5c8a629c7e6ca6d50bce0332fbb99bd683e))
+
+Full set of changes: [`4.7.6...4.7.7`][4.7.7]
+
+[4.7.7]: https://github.com/pronamic/wp-pronamic-pay-mollie/compare/v4.7.6...v4.7.7
+
 ## [4.7.6] - 2023-03-29
 ### Changed